Adding log messages for display if the user is added to the database
