--- src/main/java/org/fogbowcloud/app/api/configuration/WebConfiguration.java
@@ -1,20 +1,0 @@
-package org.fogbowcloud.app.api.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-
-@Configuration
-@EnableWebMvc
-public class WebConfiguration extends WebMvcConfigurerAdapter {
-    private static final String[] HTTP_METHODS_SUPPORTED = { "GET", "POST", "PUT", "DELETE" };
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("*").
-                allowedHeaders("*")
-                .allowedMethods(HTTP_METHODS_SUPPORTED);
-    }
-}
